+++ metagenomescope/tests/graph/test_draw_results.py
@@ -16,7 +16,7 @@ def test_init_empty():
 
 
 def test_init_nolayout_subgraph():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 20})
     sg = Subgraph(5, "sg5", [b], [], [])
     dr = DrawResults({sg: None}, [])
     assert dr.region2layout == {sg: None}
@@ -29,14 +29,14 @@ def test_init_nolayout_subgraph():
 
 
 def test_get_fancy_count_text():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 20})
     sg = Subgraph(5, "sg5", [b], [], [])
     dr = DrawResults({sg: None}, [])
     assert dr.get_fancy_count_text() == "1 node, 0 edges, 0 patterns"
 
 
 def test_repr():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 20})
     e = Edge(8, 0, 0, {})
     e2 = Edge(9, 0, 0, {"asdf": "ghjik"})
     sg = Subgraph(5, "sg5", [b], [e, e2], [])
@@ -47,7 +47,7 @@ def test_repr():
 
 
 def test_get_node_and_edge_ids():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 20})
     e = Edge(8, 0, 0, {})
     e2 = Edge(9, 0, 0, {"asdf": "ghjik"})
     sg = Subgraph(5, "sg5", [b], [e, e2], [])
@@ -59,10 +59,10 @@ def test_get_node_and_edge_ids():
 
 
 def test_add_simple():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 20})
     sg5 = Subgraph(5, "sg5", [b], [], [])
 
-    c = Node(1, "C", {"orientation": config.REV})
+    c = Node(1, "C", {"orientation": config.REV, "length": 30})
     sg6 = Subgraph(6, "sg6", [c], [], [])
 
     dr = DrawResults({sg5: None}, [])
@@ -73,10 +73,10 @@ def test_add_simple():
 
 
 def test_add_with_scope_settings():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 30})
     sg5 = Subgraph(5, "sg5", [b], [], [])
 
-    c = Node(1, "C", {"orientation": config.REV})
+    c = Node(1, "C", {"orientation": config.REV, "length": 40})
     sg6 = Subgraph(6, "sg6", [c], [], [])
 
     dr = DrawResults({sg5: None}, [ui_config.SHOW_PATTERNS])
@@ -87,7 +87,7 @@ def test_add_with_scope_settings():
 
 
 def test_add_duplicate_region():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 40})
     sg5 = Subgraph(5, "sg5", [b], [], [])
 
     dr = DrawResults({sg5: None}, [])
@@ -98,10 +98,10 @@ def test_add_duplicate_region():
 
 
 def test_add_incompatible_scope_settings():
-    b = Node(0, "B", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 40})
     sg5 = Subgraph(5, "sg5", [b], [], [])
 
-    c = Node(1, "C", {"orientation": config.REV})
+    c = Node(1, "C", {"orientation": config.REV, "length": 50})
     sg6 = Subgraph(6, "sg6", [c], [], [])
 
     dr = DrawResults({sg5: None}, [ui_config.SHOW_PATTERNS])
@@ -112,11 +112,11 @@ def test_add_incompatible_scope_settings():
 
 
 def test_get_sorted_regions_subgraphs():
-    b = Node(0, "B", {"orientation": config.FWD})
-    b2 = Node(1, "B2", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 30})
+    b2 = Node(1, "B2", {"orientation": config.FWD, "length": 30})
     sg5 = Subgraph(5, "sg5", [b, b2], [], [])
 
-    c = Node(2, "C", {"orientation": config.REV})
+    c = Node(2, "C", {"orientation": config.REV, "length": 30})
     sg6 = Subgraph(6, "sg6", [c], [], [])
 
     dr = DrawResults({sg5: None, sg6: None}, [ui_config.SHOW_PATTERNS])
@@ -125,17 +125,17 @@ def test_get_sorted_regions_subgraphs():
 
 
 def test_get_sorted_regions_subgraphs_and_components():
-    b = Node(0, "B", {"orientation": config.FWD})
-    b2 = Node(1, "B2", {"orientation": config.FWD})
+    b = Node(0, "B", {"orientation": config.FWD, "length": 30})
+    b2 = Node(1, "B2", {"orientation": config.FWD, "length": 30})
     sg5 = Subgraph(5, "sg5", [b, b2], [], [])
 
-    c = Node(2, "C", {"orientation": config.REV})
+    c = Node(2, "C", {"orientation": config.REV, "length": 30})
     sg6 = Subgraph(6, "sg6", [c], [], [])
 
-    n1 = Node(3, "1", {})
-    n2 = Node(4, "2", {})
-    n3 = Node(5, "3", {})
-    n4 = Node(6, "4", {})
+    n1 = Node(3, "1", {"length": 1})
+    n2 = Node(4, "2", {"length": 1})
+    n3 = Node(5, "3", {"length": 1})
+    n4 = Node(6, "4", {"length": 1})
     sg7 = Subgraph(7, "sg7", [n1, n2, n3, n4], [], [])
 
     dr = DrawResults(
@@ -143,7 +143,7 @@ def test_get_sorted_regions_subgraphs_and_components():
     )
     assert dr.get_sorted_regions() == [sg7, sg5, sg6]
 
-    r = Node(7, "5", {})
+    r = Node(7, "5", {"length": 3})
     cc = Component(8, [r], [], [])
 
     # even though the component only has one node, it still goes before the